CSHARP-5865: Update handshake spec to allow using OP_MSG exclusively
