+++ src/MongoDB.Driver/Core/Clusters/ClusterClock.cs
@@ -1,4 +1,4 @@
-﻿/* Copyright 2017-present MongoDB Inc.
+﻿/* Copyright 2010-present MongoDB Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -13,6 +13,7 @@
 * limitations under the License.
 */
 
+using System.Threading;
 using MongoDB.Bson;
 using MongoDB.Driver.Core.Misc;
 
@@ -49,16 +50,14 @@ namespace MongoDB.Driver.Core.Clusters
         public void AdvanceClusterTime(BsonDocument newClusterTime)
         {
             Ensure.IsNotNull(newClusterTime, nameof(newClusterTime));
-            _clusterTime = GreaterClusterTime(_clusterTime, newClusterTime);
-        }
-    }
 
-    internal sealed class NoClusterClock : IClusterClock
-    {
-        public BsonDocument ClusterTime => null;
-
-        public void AdvanceClusterTime(BsonDocument newClusterTime)
-        {
+            BsonDocument currentClusterTime;
+            BsonDocument newValue;
+            do
+            {
+                currentClusterTime = _clusterTime;
+                newValue = GreaterClusterTime(currentClusterTime, newClusterTime);
+            } while (!ReferenceEquals(Interlocked.CompareExchange(ref _clusterTime, newValue, currentClusterTime), currentClusterTime));
         }
     }
 }
